+++ R/dom.R
@@ -1,20 +1,25 @@
-#' @title Find dominated and dominant strategies in a normal-form game
+#' @title Find dominated or dominant strategies in a normal-form game
 #' @description \code{dom()} finds each player's (weakly) dominated
-#'     or (weakly) dominant strategies
-#' @param game A normal-form game object created by \code{normal_form()}.
+#'     or (weakly) dominant strategies.
+#' @return A list of dominated and weakly dominated strategies or a list of
+#'     dominant and weakly dominant strategies.
+#' @param game An object of "normal_form" class created by \code{normal_form()}.
 #'     The game's type must be "matrix".
-#' @param type Specify which to be found, "dominated" or "dominant".
-#'   Default is "dominated."
+#' @param type Which to be found, "dominated" or "dominant".
+#'     Default is \code{"dominated"}.
+#' @param quietly If \code{TRUE}, the message telling dominated (or dominant)
+#'     strategies will not be shown. Default is \code{FALSE}.
 #' @author Yoshio Kamijo and Yuki Yanai <yanai.yuki@@kochi-tech.ac.jp>
 #' @export
 dom <- function(game,
-                type = "dominated") {
+                type = "dominated",
+                quietly = FALSE) {
 
-  match.arg(type, choices = c("dominated", "dominant"))
+  type <- match.arg(type, choices = c("dominated", "dominant"))
 
   if (type == "dominated") {
-    find_dominated(game)
+    find_dominated(game, quietly = quietly)
   } else {
-    find_dominant(game)
+    find_dominant(game, quietly = quietly)
   }
 }