+++ R/find_mixed_NE.R
@@ -1,9 +1,9 @@
 #' @title Find a mixed-strategy Nash equilibrium.
-#' @description \code{find_mixed_NE} finds a mixed-strategy Nash equilibrium of
+#' @description \code{find_mixed_NE} finds mixed-strategy Nash equilibrium of
 #'     a normal-form game with discrete-choice strategies.
 #' @return A list of the probabilities given to each strategy that specifies
 #'     the mixed-strategy Nash equilibrium and a list of mixed-strategy Nash
-#'     equalibrium of the subsets of strategies..
+#'     equilibria of the subsets of strategies
 #' @param game A "normal_form" class object created by \code{normal_form()}.
 #'     The game's type must be "matrix".
 #' @seealso \code{\link{normal_form}}
@@ -34,7 +34,7 @@ find_mixed_NE <- function(game) {
                       col = 1:length(s2_sets))
 
   msNE_list <- NULL
-
+  n_msNE <- 0
   for (k in 1:nrow(pair)) {
 
     s1_sub <- s1_sets[[pair$row[k]]]
@@ -84,13 +84,16 @@ find_mixed_NE <- function(game) {
           !all(prob2 >= 0) | !all(prob2 <=1)) {
         msNE <- NULL
       } else {
+        n_msNE <- n_msNE + 1
         msNE <- list(s1 = prob1, s2 = prob2)
+
+        msNE_list[[n_msNE]] <- list(s1 = s1_sub,
+                                s2 = s2_sub,
+                                msNE = msNE)
       }
     }
-    msNE_list[[k]] <- list(s1 = s1_sub,
-                           s2 = s2_sub,
-                           msNE = msNE)
   }
 
-  return(list(msNE = msNE, msNE_list = msNE_list))
+  return(list(msNE = msNE,
+              msNE_list = msNE_list))
 }